ppcfd: migrate models of ppdiffusion to ppcfd
