--- examples/ppdiffusion/models/__init__.py
@@ -1,12 +1,0 @@
-from __future__ import annotations
-
-
-from .base_model import BaseModel  # isort:skip
-from .unet_simple import UNet as SimpleUnet  # isort:skip
-from .modules.ema import LitEma  # isort:skip
-
-__all__ = [
-    "BaseModel",
-    "SimpleUnet",
-    "LitEma",
-]
