+++ ppcfd/models/ppdiffusion/modules/__init__.py
@@ -1,0 +1,32 @@
+# Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from ppcfd.models.ppdiffusion.modules.ema import LitEma
+from ppcfd.models.ppdiffusion.modules.misc import LearnedSinusoidalPosEmb
+from ppcfd.models.ppdiffusion.modules.misc import Residual
+from ppcfd.models.ppdiffusion.modules.misc import SinusoidalPosEmb
+from ppcfd.models.ppdiffusion.modules.misc import default
+from ppcfd.models.ppdiffusion.modules.misc import get_normalization_layer
+from ppcfd.models.ppdiffusion.modules.misc import get_time_embedder
+
+
+__all__ = [
+    "LitEma",
+    "default",
+    "get_time_embedder",
+    "get_normalization_layer",
+    "Residual",
+    "SinusoidalPosEmb",
+    "LearnedSinusoidalPosEmb",
+]
