+++ ppcfd/models/ppdiffusion/unet_simple.py
@@ -1,8 +1,22 @@
+# Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import paddle
 from einops import rearrange
 
-from .base_model import BaseModel
-from .modules.misc import get_time_embedder
+from ppcfd.models.ppdiffusion import BaseModel
+from ppcfd.models.ppdiffusion.modules import get_time_embedder
 
 
 RELU_LEAK = 0.2
